src/requirements.txt
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.12.1
attrs==26.1.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.1.8
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.1
fastapi==0.128.8
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.18.1
numpy==2.0.2
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.6
pyarrow==21.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
starlette==0.49.3
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.39.0
